measurement converter v2
- added volume measurements
- completely revamped conversion choice process to improve user experience
- shortened the code by about half

I wanted to add the possibility of choosing both units to convert independently, while I was adding that I thought of a way to shorten a lot of the code that stores the conversions. I decided to leave that for later, but then an error came up and I realized it could be fixed by implementing the other system, so I did.

As mentioned, you can now choose the two units you'd like to convert between independently of one another. Each unit is assigned a number, these are concatenated to form a key to get the conversion value. The conversions now all use the same formula but a different value, which means I don't have to define each one seperately anymore, just have a dictionary of the different values.